Split the remotetool into a separate library
In this CL, I'm creating an embeddedtool binary so that I can
subsequently use it in reproxytool binary in re-client to implement the
log file filtering functionalities. There's no behavior change
associated with this CL.

Tested: Ran show_action and show_action_2 (error'ed out) on the new
remotetool.
