+++ go/pkg/tool/BUILD.bazel
@@ -2,7 +2,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "tool",
-    srcs = ["tool.go"],
+    srcs = [
+        "embeddedtool.go",
+        "tool.go",
+    ],
     importpath = "github.com/bazelbuild/remote-apis-sdks/go/pkg/tool",
     visibility = ["//visibility:public"],
     deps = [
